Use one root eslint config
