+++ eslint.config.mjs
@@ -1,17 +1,44 @@
-const config = {
-  ignores: [
-    "/node_modules",
-    "/.pnp",
-    ".pnp.js",
-    ".yarn/install-state.gz",
-    "/website/.next/",
-    "/editor/.next/",
+// @ts-check
 
-    "/website/out/",
+import nextPlugin from "@next/eslint-plugin-next";
+import reactPlugin from "eslint-plugin-react";
+import hooksPlugin from "eslint-plugin-react-hooks";
+import eslint from "@eslint/js";
+import tseslint from "typescript-eslint";
+import { fixupPluginRules } from "@eslint/compat";
 
-    "/content",
-    "/test-content",
-  ],
-};
+const patchedNextPlugin = fixupPluginRules(nextPlugin);
+const patchedHooksPlugin = fixupPluginRules(hooksPlugin);
 
-export default config;
+export default [
+  eslint.configs.recommended,
+  ...tseslint.configs.recommended,
+  {
+    files: ["**/*.ts", "**/*.tsx"],
+    plugins: {
+      react: reactPlugin,
+      "react-hooks": patchedHooksPlugin,
+      "@next/next": patchedNextPlugin,
+    },
+    rules: {
+      ...reactPlugin.configs["jsx-runtime"].rules,
+      ...hooksPlugin.configs.recommended.rules,
+      ...nextPlugin.configs.recommended.rules,
+      ...nextPlugin.configs["core-web-vitals"].rules,
+      "@next/next/no-img-element": "error",
+    },
+    ignores: [
+      "/node_modules",
+      "/.pnp",
+      ".pnp.js",
+      ".yarn/install-state.gz",
+      "/website/.next/",
+      "/editor/.next/",
+
+      "/website/out/",
+
+      "/content",
+      "/test-content",
+    ],
+  },
+];
