--- websites/recipe-website/export/eslint.config.mjs
@@ -1,21 +1,0 @@
-import nextPlugin from "@next/eslint-plugin-next";
-import reactPlugin from "eslint-plugin-react";
-import hooksPlugin from "eslint-plugin-react-hooks";
-
-export default [
-  {
-    files: ["**/*.ts", "**/*.tsx"],
-    plugins: {
-      react: reactPlugin,
-      "react-hooks": hooksPlugin,
-      "@next/next": nextPlugin,
-    },
-    rules: {
-      ...reactPlugin.configs["jsx-runtime"].rules,
-      ...hooksPlugin.configs.recommended.rules,
-      ...nextPlugin.configs.recommended.rules,
-      ...nextPlugin.configs["core-web-vitals"].rules,
-      "@next/next/no-img-element": "error",
-    },
-  },
-];
